rapido: initial release
As documented in readme.txt, Rapido is a utility for rapid testing of
Linux kernel changes. It was initially targeted at Ceph RBD and CephFS
kernel client functionality, but now also covers a number of standalone
use cases.

The scripts that make up Rapido are in themselves quite brainless. Most
of the heavy lifting is instead performed by:
- Dracut
  + Generates a VM image, with kernel-modules and minimal user-space.
- QEMU / KVM
  + Boots a given Dracut VM image and compiled Linux kernel on the local
    system.

Signed-off-by: David Disseldorp <ddiss@suse.de>