runtime.vars: add helper to check fstests configuration
_rt_require_fstests() checks that the fstests source directory exists,
and that the check executable is present.

This prepares for migration from a FSTESTS_DIR=<parent_path> rapido.conf
parameter, where <parent_path> is assumed to have an immediate xfstests
subdir, to an FSTESTS_SRC=<src_path>. This migration is possible now
that Dracut correctly handles installation of immediate files, when
installed via "dracut --include <dir_path> <initramfs_path>" - this was
broken in previous versions.

Signed-off-by: David Disseldorp <ddiss@suse.de>