blktests: add cut and autorun scripts
These scripts can be used to generate a minimal initramfs VM image for
blktests execution. A 1G zram device is provisioned and configured as a
test target device.

The new BLKTESTS_SRC and BLKTESTS_AUTORUN_CMD rapido.conf parameters
are added, for definition of the test-suite source and invocation
command respectively.

Signed-off-by: David Disseldorp <ddiss@suse.de>