Add support for LTP
The Linux Test Project [1] is a collection of tools for testing the
Linux kernel and related features.
These tests can/will harm the system you run them on, so running them
inside virtual machines is neccessary for developing new testcases.

Rapido can build a ramdisk image with the complete LTP binaries and
execute a single test or the included testrunner.

[1] https://github.com/linux-test-project/ltp/

Signed-off-by: Michael Moese <mmoese@suse.de>