cut/qemu_rbd: fix init loop
qemu-rbd has always been special in that it doesn't have a corresponding
autorun script. The hack to support this in the past was to put
vm_autorun.env at .profile, but with 3112ea3f2721454057e085d3eeaf47464dd
this turned into an init loop.
_rt_require_dracut_args() and vm_autorun.env now support booting without
an autorun script, so use this.

Signed-off-by: David Disseldorp <ddiss@suse.de>