vm: use net-conf/vm# directories to configure network
This works quite nicely and is quite expressive - tap devices are
assigned to VMs based on the .network prefix in the corresponding vm#
directory, allowing for multiple NICs per VM.

MAC addresses (for netd [match]) are generated by taking an md5sum hash
of the vm#.tapdev string.

The removed net-conf/*.network examples will be replaced with a separate
net-conf.example directory in future.

Signed-off-by: David Disseldorp <ddiss@suse.de>