runtime: add and load virtio-rng / 9p modules on demand
If the virtio-rng and 9p virtfs modules aren't built-in (e.g. when
booting a distro kernel with KERNEL_SRC unset), then the modules need
to be explicitly added to the initramfs image and loaded on boot.

Do so if the corresponding virtio-rng or 9p functionality is requested
via QEMU_EXTRA_ARGS and / or VIRTFS_SHARE_PATH respectively.

Signed-off-by: David Disseldorp <ddiss@suse.de>