samba-local: configure Btrfs backed shares
Users were able to select XFS or Btrfs as the Samba share's backing
filesystem via the filesystem variable in the autorun script, but this
adds unnecessary ugliness. Remove the conditional fs logic and just
provision Btrfs.
In future I expect that we'll carry both samba-btrfs and samba-xfs
runners, with samba-local made an alias for the former.

One other minor change in this commit is the use of FSTESTS_ZRAM_SIZE
when determining the zram size - this runner is frequently used as a
server for fstests-cifs, so I think it makes sense, rather than a
static 1G value.

Signed-off-by: David Disseldorp <ddiss@suse.de>