runtime: don't fail if default rapido.conf is missing
Provide a warning if $RAPIDO_DIR/rapido.conf is missing, instead of
exiting via _fail(). The aim here is to make it as easy and quick as
possible to go from a clone to a booted rapido VM (reusing the host
kernel).
If the RAPIDO_CONF environment variable is set prior to rapido
invocation then continue to fail if the explicitly provided rapido.conf
path is missing.

We already have checks in most places to ensure required variables are
set, so not much more is needed. The only real change is ensuring that
QEMU_EXTRA_ARGS="-nographic -device virtio-rng-pci" defaults are set
outside of rapido.conf, as QEMU will otherwise default to a graphical
UI.

Signed-off-by: David Disseldorp <ddiss@suse.de>