vm_autorun.env: make script idempotent
vm_autorun.env is copied into the rapido root fs as /etc/profile,
and as such will be executed every time a login shell is invoked
(ie via 'su -' or 'sh -l'). As this script also serves as system
initialisation we need to make sure to run the initialisation only
once, and only export the environment variables for later invocations.

Signed-off-by: Hannes Reinecke <hare@suse.de>