usbip: server and client cut / autorun scripts
The usbip-vudc-zram provisions a zram device and exposes it over the
network via usbipd, which needs to be present under KERNEL_SRC.

The client carries the usbip binary and just prints instructions for
connecting to a server. A USBIP_SERVER rapido.conf parameter which is
auto-connected might make sense in future.

Signed-off-by: David Disseldorp <ddiss@suse.de>