vm_autorun.env: Provide shutdown handler with systemd
There are cuts that setup the environment with fully-fledged PID1 as
systemd. Pulling the plug in such cases may or may not work depending on
the ACPI support in the kernel/hypervisor.
The system may be left in not well-defined state when _only_ sysrq is
invoked where most of the userspace is left running.

Fix this by detecting which kind of setup we're in and ask PID1 to
nicely shut down (via signals to avoid dependency on systemctl).