selftest: fix Dracut log levels for Tumbleweed
Tumbleweed is carrying the following Dracut change:

  commit 311f84f698fb333fa80af4fe6b2a78304b4364cb
  ...
  ci(suse.conf.example): change log levels

  Reduce default standard log verbosity from `info` to `warning`, and increase
  syslog verbosity from none to `info`.

This breaks selftest, which uses expect to detect specific log messages.
This appears to be all that's needed to get things working again.
The tests should probably be fixed to avoid relying on any Dracut
output, but let's go with a config change for now.

Link: https://github.com/rapido-linux/rapido/issues/254
Signed-off-by: David Disseldorp <ddiss@suse.de>