rapido-cut: drop "--install src→dst" syntax
This syntax was always very cumbersome, and we only really added it for
installing /init (from a rapido or systemd source). We now handle /init
installation within rapido-cut itself.
If we do bring something like this back then we should definitely go
with a syntax similar to "--include src dst" instead.

Signed-off-by: David Disseldorp <ddiss@suse.de>