rapido-cut: search /lib, /lib64, /bin and /sbin
The current bin and lib search paths assume a usr-merge system, which
results in missing libraries on non-usr-merge systems such as Leap 15.6.

Change missing tracking to actually return an error, so that users are
aware of the search failure. In future we could add best-effort
searching via --try-install or something.

Signed-off-by: David Disseldorp <ddiss@suse.de>