rapido-cut: add --try-install parameter
rapido-cut --install now aborts on missing bin/lib items. xfstests and
other test suites have a bunch of soft dependencies, where certain tests
will only run if a specific binary is available.

--try-install will archive a binary if available, or skip it if missing.
ELF dependencies for a --try-install binary are handled as mandatory: if
a binary is installed then it's reasonable to assume that the package
manager has ensured that all dependencies are available.

Signed-off-by: David Disseldorp <ddiss@suse.de>