rapido-cut: use ELF RUNPATH paths for library lookup
Parse DT_RUNPATH alongside DT_NEEDED and ensure that any provided paths
are used alongside the default library search paths.

As an example, fstests cut scripts install fio, which depends on
librados.so. librados.so requires libceph-common.so, which lives under
/usr/lib64/ceph/ on Leap 15.6 and Tumbleweed. /usr/lib64/ceph/ is
carried as a DT_RUNPATH entry for librados.so. This change ensures that
libceph-common.so is successfully located.

We should also be able to remove the systemd library search path from
defaults, as it appears that systemd also uses DT_RUNPATH.

Signed-off-by: David Disseldorp <ddiss@suse.de>