cut/fstests_btrfs: use --try-install for soft dependencies
xfstests has plenty of soft dependencies, where a test may be run if a
binary (kmod, option, etc.) exists or otherwise skipped. Move a bunch
of these binaries from --install to --try-install so that we don't abort
if these soft dependencies are missing.

One other change here is that we now pull in nano instead of vim. vim
may be a symlink to gvim, which pulls in a huge amount of libraries.
nano should be a little bit lighter.

Signed-off-by: David Disseldorp <ddiss@suse.de>