kv-conf: rework kv_process to return kv tuple
Pass the mutable line buffer to kv_process, so that it can clear any
processed data and retain unprocessed multi-line portions. A (now
non-mut) map still needs to be provided for ${} variable substitution.

This should put us in a better position to support callback based kv
parsing or (perhaps?) iter().

Signed-off-by: David Disseldorp <ddiss@suse.de>